% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_match.R
\name{as_tibble.cqc_match_result}
\alias{as_tibble.cqc_match_result}
\title{Convert the match_result to table}
\usage{
\method{as_tibble}{cqc_match_result}(x)
}
\arguments{
\item{x}{"cqc_match_result" returned as part of cqc_match call}
}
\description{
Convert the match_result to table
}
