% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_io.R
\name{cqc_cf_list}
\alias{cqc_cf_list}
\title{Construct a \code{cqc_cf_list} object from a list of \code{cytoframe} objects}
\usage{
cqc_cf_list(x)
}
\arguments{
\item{x}{a named list of \code{\link[flowWorkspace]{cytoframe}} objects}
}
\description{
This is the core data object for \code{\link[cytoqc:cytoqc-package]{cytoqc}}.
}
\examples{
# This is just for illustration. cqc_load_fcs will normally take care of this step.
fcs_files <- list.files(system.file("extdata", "GvHD_QC", package = "cytoqc"), full.names = TRUE)
cf_list <- lapply(fcs_files[1:3], load_cytoframe_from_fcs)
names(cf_list) <- fcs_files[1:3]

# Construct a cqc_cf_list object from a list of cytoframes
cf_list <- cqc_cf_list(cf_list)

}
