% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_cluster.R, R/cqc_plot.R
\name{cqc_cluster}
\alias{cqc_cluster}
\alias{cqc_plot.cqc_cluster}
\title{Cluster groups based on the result of cqc_check}
\usage{
cqc_cluster(x, ...)

\method{cqc_plot}{cqc_cluster}(y, ...)
}
\arguments{
\item{x}{object resulting from a call to \code{\link{cqc_check}}}

\item{...}{ngroup: integer specifying number of groups in to which the clusters will be combined

height: number between 0 and 1 that specifies a cut height for the tree to determine the groups. Ignored if ngroup is specified.

missing_penalty: multiplicative factor used to impose penalty for labels missing from a group. The distance for a missing label
will be calculated as missing_penalty times the largest distance between non-missing labels. Defaults to 1.}

\item{y}{an object of class cqc_cluster resulting from a cqc_cluster call}
}
\value{
a \code{cqc_cluster} object whose primary value is the tibble in its \code{group_membership} slot providing the map between the original
groups and the combined groups resulting from clustering.
}
\description{
Sets of labels (strings representing channels, markers, keywords, or panels consisting of channel/marker pairs) for each group
will be clustered based on the summed distances of their aligned elements. This provides an automated way to look for groups with
sets of labels that likely represent the same underlying values but with minor differences or errors.
}
