% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_fix.R
\name{cqc_delete}
\alias{cqc_delete}
\title{Delete methods for cyto data}
\usage{
cqc_delete(x, value, type, ...)
}
\arguments{
\item{x}{cytoframe or GatingSet}

\item{value}{the value to be deleted}

\item{type}{one of qc task "channel", "marker", "keyword", "gate"}

\item{...}{unused}
}
\description{
It is typically called automatically by cqc_fix call
}
