% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_fix.R
\name{cqc_drop_groups}
\alias{cqc_drop_groups}
\title{Helper function to remove the outlier groups that can't be fixed}
\usage{
cqc_drop_groups(cqc_groups, id)
}
\arguments{
\item{cqc_groups}{the object returned by 'cqc_check'}

\item{id}{the group id to be dropped from the dataset}
}
\description{
Helper function to remove the outlier groups that can't be fixed
}
