% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_params.R
\name{cqc_find_solution.cqc_report}
\alias{cqc_find_solution.cqc_report}
\title{cqc_find_solution}
\usage{
\method{cqc_find_solution}{cqc_report}(x, max.distance = 0.1)
}
\arguments{
\item{max.distance}{Maximum distance allowed for a match. See ?agrep}
}
\description{
cqc_find_solution
}
