% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_params.R
\name{cqc_fix.cqc_solution}
\alias{cqc_fix.cqc_solution}
\title{Apply the cqc_solution}
\usage{
\method{cqc_fix}{cqc_solution}(x, func)
}
\arguments{
\item{x}{the cqc_solution returned by 'find_solution' calls}
}
\description{
Peform the actual fixing action (i.e update or delete)
}
