% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_fix.R
\name{cqc_fix}
\alias{cqc_fix}
\title{Apply the cqc_solution}
\usage{
cqc_fix(x, ...)
}
\arguments{
\item{x}{the cqc_solution returned by 'find_solution' calls}

\item{...}{addiitional arguments not for the user.}
}
\description{
Peform the actual fixing action (i.e update or delete)
}
