% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_group.R
\name{cqc_group.cqc_data}
\alias{cqc_group.cqc_data}
\title{QC check}
\usage{
\method{cqc_group}{cqc_data}(
  x,
  type = c("channel", "marker", "panel", "keyword"),
  delimiter = "|"
)
}
\arguments{
\item{x}{cqc_data}

\item{type}{specify the qc type, can be "channel", "marker" or "panel"}

\item{delimiter}{a special character used to separate channel and marker}
}
\value{
a tibble with 4 columns: FCS, qc type (e.g. channel), group_id and nFCS (i.e. group count)
}
\description{
This is the first step of the entire qc workflow.
It extracts meta information(specified by 'type' argument) from the raw data
and compare/group them across samples.
This provides a sample-wise data table for the further summary report.
}
\examples{
\dontrun{
groups <- cqc_group(cqc_data, "channel")
}
}
