% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_group.R
\name{cqc_group}
\alias{cqc_group}
\title{Perform a QC check on flow data.}
\usage{
cqc_group(x, ...)
}
\arguments{
\item{x}{cqc_cf_list}

\item{...}{additional arguments.
type specify the qc type, can be "channel", "marker" or "panel"
delimiter a special character used to separate channel and marker
keys The vector to supply the keys to be grouped on. default is NULL, which is extracted automatically from the flow data}
}
\value{
a tibble with 4 columns: object, qc type (e.g. channel), group_id and nobject (i.e. group count)
}
\description{
This is the first step of the entire qc workflow.
It extracts meta information(specified by 'type' argument) from the raw data
and compare/group them across samples.
This provides a sample-wise data table for the further summary report.
}
\examples{
\dontrun{
groups <- cqc_group(cqc_cf_list, "channel")
}
}
