% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_io.R
\name{cqc_gs}
\alias{cqc_gs}
\title{Construct a 'cqc_gs' object from a 'GatingSet'}
\usage{
cqc_gs(x)
}
\arguments{
\item{x}{a GatingSet object}
}
\description{
This is mainly for distpatch and wrapping many of the
qc operations on a \code{\link{cqc_cf_list}} object for the GatingSet's
underlying data
}
