% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_populations.R
\name{cqc_gs_check_populations}
\alias{cqc_gs_check_populations}
\title{post-gating qc check}
\usage{
cqc_gs_check_populations(gs, nodes = NULL, plot = FALSE)
}
\arguments{
\item{gs}{GatingSet}

\item{nodes}{the populations to check}

\item{plot}{whether to plot the gates}
}
\description{
post-gating qc check
}
