% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_io.R
\name{cqc_gs_list}
\alias{cqc_gs_list}
\title{Construct a 'cqc_gs_list' object from a list of 'GatingSet' objects}
\usage{
cqc_gs_list(x)
}
\arguments{
\item{x}{a list of 'GatingSet' objects}
}
\description{
For the methods dispatching purpose
}
