% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_fix.R
\name{cqc_insert}
\alias{cqc_insert}
\title{insertion methods for cyto data}
\usage{
cqc_insert(x, value, type, ...)
}
\arguments{
\item{x}{cytoframe or GatingSet}

\item{value}{the value to be inserted}

\item{type}{one of qc task only  "keyword" is valid for now.}

\item{...}{unused}
}
\description{
It is typically called automatically by cqc_fix call
}
