% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_match_reference.R
\name{cqc_match_reference}
\alias{cqc_match_reference}
\title{find the the difference between the reference and target group}
\usage{
cqc_match_reference(x, ...)
}
\arguments{
\item{x}{cqc_group object returned by cqc_group call}

\item{...}{ref specifies the reference, which can be either an integer group id or a characte vector giving the actual values of the reference
select the group ids selected for processing
type the qc type (either "channle", "marker", "gate")}
}
\description{
find the the difference between the reference and target group
}
