% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_plot.R
\name{cqc_plot}
\alias{cqc_plot}
\title{Specialized plots for objects from the \code{cytoqc} package}
\usage{
cqc_plot(y, ...)
}
\arguments{
\item{y}{object to be plotted}

\item{...}{additional arguments (currently unused)}
}
\description{
This is a generic function for plotting functionality specific
to objects from the \code{\link[=cytoqc-package]{cytoqc}} package.
}
\details{
It is currently only implemented for results of \code{\link{cqc_cluster}}.
}
