% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_find_solution.R
\name{cqc_recommend}
\alias{cqc_recommend}
\alias{cqc_recommend.cqc_match_result}
\title{Find solution to resolve the discrepancy discovered by match_reference}
\usage{
\method{cqc_recommend}{cqc_match_result}(x, max.distance = 0.1, ...)
}
\arguments{
\item{x}{A CQC object of some kind. See vignettes.}

\item{max.distance}{Maximum distance allowed for a match. See ?agrep}

\item{...}{additional arguments not for the user.}
}
\value{
a table (with 'from' and 'to' columns) represents the itemized fix recommendation. When 'to' is 'NA', it means the entry is redundunt and can be removed
}
\description{
It tries to find the aproximate match(based on 'agrep') between the target and reference as well as the extra redundunt items that can be removed.
}
\examples{
\dontrun{
solution <- cqc_recommend(groups, select = c(1, 4))
}
}
