% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_fix.R
\name{cqc_set_panel}
\alias{cqc_set_panel}
\title{Update panel info}
\usage{
cqc_set_panel(x, panel, ref.col, ...)
}
\arguments{
\item{x}{cqc_data}

\item{panel}{a tibble contains channel and marker colums (typically returned by 'cf_get_panel` )}

\item{ref.col}{the column used as the reference, either 'channel' or 'marker'}
}
\description{
It updates either 'channel' or 'marker' info depends on the value of 'ref.col'
if ref.col is set to 'channel', then 'channel' column from 'panel' is used as the
reference to be matched against 'cqc_data' and 'marker'  will be set according to 'panel'
}
