% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_match.R
\name{cqc_update_match}
\alias{cqc_update_match}
\alias{cqc_delete_match}
\title{Manual update/delete the match report}
\usage{
cqc_update_match(x, map, group_id = NULL)

cqc_delete_match(x, map, group_id = NULL)
}
\arguments{
\item{x}{'cqc_match_result_and_solution' object generated by "cqc_match"}

\item{map}{For update, it is a named vector that provides the paired strings(i.e. old value vs new value).
For deletion, it is a character vector that provides the items to be deleted from matched pairs.}

\item{group_id}{the group to update. If NULL, then operate on all groups}
}
\description{
There are cases that the automatic string match from `cqc_match` can't resolve.
This function provides the remedy to manually update the match result.
}
\examples{
fcs_files <- list.files(system.file("extdata", "GvHD_QC", package = "cytoqc"), full.names = TRUE)
qc_cf_list <- cqc_load_fcs(fcs_files)
groups <- cqc_check(qc_cf_list, type = "marker")
match_result <- cqc_match(groups, ref = 3)

# add the matching pairs to the auto-matching results
match_result <- cqc_update_match(match_result, map = c("PTPRC PE" = "CD45 PE"), group_id = 5)

# You can delete any match added manually or automatically by cqc_match
match_result <- cqc_delete_match(match_result, "PTPRC PE", group_id = 5)

# You can also omit the group_id to apply to all groups
match_result <- cqc_update_match(match_result, map = c("PTPRC PE" = "CD45 PE"))
}
