% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_io.R
\name{cqc_write_fcs}
\alias{cqc_write_fcs}
\title{Write out tidied flow data (cqc_cf_list) back to fcs}
\usage{
cqc_write_fcs(x, out, verbose = TRUE, ...)
}
\arguments{
\item{x}{cqc_cf_list}

\item{out}{the output directory that the FCS will be written}

\item{verbose}{whether to print each sample name during the writing process}

\item{...}{other arguments passed down to 'write.FCS'}
}
\description{
Write out tidied flow data (cqc_cf_list) back to fcs
}
