% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoqc-package.R
\docType{package}
\name{cytoqc-package}
\alias{cytoqc}
\alias{cytoqc-package}
\title{cytoqc: Quality control and standardization of cytometry data}
\description{
Provides quality control, quality assessment, and data standardization tools for cytometry.
}
\details{
The cytoqc package is used to perform QC and data standardization for flow and mass cytometry data stored in
GatingSet objects or FCS files.

It simplifies the tasks of standardizing channel, markers, keywords, gating schemes, gate names, and so forth, necessary to
run cytometry data through automated analysis pipelines.

The process for running cytoqc from FCS files consists of several steps.

First, one loads the data from FCS files using  \code{cqc_load_fcs()}.
The data is loaded into a \code{cqc_cf_list} object.

Next a QC evaluation of  "channel", "marker" or "panel" is performed using \code{cqc_check()}.
This returns a \code{cqc_check} family object, specific to either the "channel", "marker" or "panel" qc.
This object can be viewed using \code{summary}.
A nice print out of the \code{cqc_check} for reports can be generated with \code{knit_print}.

Next a reference group is chosen using \code{cqc_match()} and passing in a vector of channels or markers to use as the referece.
Alterately one can pass in the group_id of the group to use as a reference.
This returns a \code{cqc_match_result}.

Next we resolve discrepancies using \code{cqc_recommend} on the \code{cqc_match_result}.
This will propose several ways to resolve discrepancies in the markers, channels, or panel.
Again, a pretty print out can be generated using \code{knit_print}.

Finally the proppsed solution can either be edited (by writing it out to a csv file and editing it), or by applying \code{cqc_fix()} to the \code{cqc_solution}.

The QC can be updated after applying the fix using \code{cqc_check} again.

Any groups that cannot be standardized can be dropped using \code{cqc_drop_groups}.

The tidied data can be coerced to a \code{cytoset} via a call to the \code{cytoset()} constructor.

To summarize:
Read -> group -> set reference -> propose a solution to qc issues -> apply the solution -> store cleaned data.

See the different vignettes for additional details.
}
\keyword{internal}
