% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_group.R
\name{diff.cqc_check}
\alias{diff.cqc_check}
\title{Display the differences among QC groups}
\usage{
\method{diff}{cqc_check}(x, vars, ...)
}
\arguments{
\item{x}{the grouped summary report generated by 'summary' call on the 'cqc_check' results}

\item{vars}{variable to split by. Determined automatically.}

\item{...}{Additional arguments not for the user. Ignore.}
}
\description{
Display the differences among QC groups
}
\examples{
\dontrun{
su <- summary(groups)
diff(su)
}
}
