% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_params.R
\name{fs_qc_cellcount}
\alias{fs_qc_cellcount}
\title{QA processes of cellcount}
\usage{
fs_qc_cellcount(fs, grouping = NULL, func = qoutlier, ...)
}
\value{
a s3 object that contains the sample names, cell count and outliers detected.
This object can be printed or plotted by its dedicated s3 method.
}
\description{
detect aberations in the number of events per sample.
If there is a natural grouping among the samples, this can be
specified using the \code{grouping} argument. In this case, the
outlier detection will be performed within its respective group for a
particular sample.
}
\examples{
data(GvHD)
fs <- GvHD[1:10]
res <- fs_qc_cellcount(fs)
}
