% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.cqc_check_summary}
\alias{knit_print.cqc_check_summary}
\title{Customized knit print for cqc_check_summary}
\usage{
\method{knit_print}{cqc_check_summary}(x, collapse = TRUE, ...)
}
\arguments{
\item{x}{cqc_check_summary object returned by 'summary' call on `cqc_check`}

\item{collapse}{whether to collapse the same information within each group}

\item{...}{not used}
}
\description{
Customized knit print for cqc_check_summary
}
