% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{match_result_to_dt}
\alias{match_result_to_dt}
\title{This can not be called at rstudio console since knit_print requires PhantomJS when run interactively}
\usage{
match_result_to_dt(x, ...)
}
\description{
This can not be called at rstudio console since knit_print requires PhantomJS when run interactively
}
