% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_group.R
\name{split.cqc_check}
\alias{split.cqc_check}
\title{Split the result of 'cqc_check' into groups}
\usage{
\method{split}{cqc_check}(x, f, drop = FALSE, ...)
}
\arguments{
\item{x}{cqc_check object}

\item{f, drop, ...}{not used}
}
\description{
It is used to split samples into separate groups when they can't be reconciled into the sampe group.
}
