% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_cluster.R
\name{split.cqc_cluster}
\alias{split.cqc_cluster}
\title{Split the result of 'cqc_cluster' into groups}
\usage{
\method{split}{cqc_cluster}(x, f, drop = FALSE, ...)
}
\arguments{
\item{x}{cqc_cluster object}

\item{f, drop, ...}{not used}
}
\description{
It is used to split samples into separate groups when they can't be reconciled into the same group.
}
