% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_group.R
\name{summary.cqc_check}
\alias{summary.cqc_check}
\title{Provide the summary view of the cqc_check report.}
\usage{
\method{summary}{cqc_check}(object, ...)
}
\arguments{
\item{object}{qc table returned by 'cqc_check'}

\item{...}{Additional arguments not for the user. Ignore.}
}
\description{
It summarise the sample-wise qc report into group overview report for reference selection and further QC action.
}
\examples{
\dontrun{
su <- summary(groups)
}
}
