% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_params.R
\name{summary.cqc_group}
\alias{summary.cqc_group}
\title{Provide the summary view of the qc report}
\usage{
\method{summary}{cqc_group}(object)
}
\arguments{
\item{object}{qc table returned by 'cqc_check'}
}
\description{
It summarise the sample-wise qc report into group overview report for reference selection and further QC action.
}
\examples{
\dontrun{
su <- summary(groups)

}
}
