% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_params.R
\name{write_fcs.cqc_data}
\alias{write_fcs.cqc_data}
\title{The helper function to write the cleaned cqc_data back to fcs}
\usage{
\method{write_fcs}{cqc_data}(x, out, verbose = TRUE)
}
\arguments{
\item{x}{cqc_data}

\item{out}{the output directory that the FCS will be written}
}
\description{
The helper function to write the cleaned cqc_data back to fcs
}
