% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_io.R
\name{write_fcs}
\alias{write_fcs}
\title{Write out tidied flow data (cqc_cf_list) back to fcs}
\usage{
write_fcs(x, ...)
}
\arguments{
\item{x}{cqc_cf_list}

\item{...}{additional arguments.
out the output directory that the FCS will be written
verbose whether to print each sample name during the writing process}
}
\description{
Write out tidied flow data (cqc_cf_list) back to fcs
}
