% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_check.R
\name{cf_get_panel}
\alias{cf_get_panel}
\title{Extract channel/marker info from a cytoframe object}
\usage{
cf_get_panel(cf, skip_na = FALSE)
}
\arguments{
\item{cf}{cytoframe}

\item{skip_na}{whether to skip the entries with marker unset (i.e. NA)}
}
\value{
a tibble with two columns: "channel" and 'marker'
}
\description{
Extract channel/marker info from a cytoframe object
}
\examples{
fcs_path <- system.file("extdata", "GvHD_QC", "s5a01.fcs", package = "cytoqc")
cf <- load_cytoframe_from_fcs(fcs_path)
cf_get_panel(cf)
}
