% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_io.R
\name{cqc_cf_list}
\alias{cqc_cf_list}
\title{Construct a 'cqc_cf_list' object from a list of 'cytoframe' objects}
\usage{
cqc_cf_list(x)
}
\arguments{
\item{x}{a list of 'cytoframe' objects}
}
\description{
This is the core data object for CytoQC.
}
