% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_check.R
\name{cqc_drop_groups}
\alias{cqc_drop_groups}
\title{Remove outlier groups from the result of 'cqc_check'.}
\usage{
cqc_drop_groups(groups, id)
}
\arguments{
\item{groups}{the object returned by 'cqc_checks'}

\item{id}{the group id to be dropped from the dataset}
}
\description{
Remove outlier groups from the result of 'cqc_check'.
}
\examples{
fcs_files <- list.files(system.file("extdata", "GvHD_QC", package = "cytoqc"), full.names = TRUE)
qc_cf_list <- cqc_load_fcs(fcs_files)
channel_groups <- cqc_check(qc_cf_list, type = "channel")
summary(channel_groups)
channel_groups <- cqc_drop_groups(channel_groups, 2)
channel_groups
}
