% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_group.R
\name{cqc_drop_groups}
\alias{cqc_drop_groups}
\title{Remove outlier groups from the result of 'cqc_prepare'.}
\usage{
cqc_drop_groups(groups, id)
}
\arguments{
\item{groups}{the object returned by 'cqc_prepares'}

\item{id}{the group id to be dropped from the dataset}
}
\description{
Remove outlier groups from the result of 'cqc_prepare'.
}
