% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_fix.R
\name{cqc_fix}
\alias{cqc_fix}
\title{Apply the cqc_solution}
\usage{
cqc_fix(x, ...)
}
\arguments{
\item{x}{the \code{cqc_solution} returned by \code{\link{cqc_match}} calls}

\item{...}{addiitional arguments not for the user.}
}
\description{
Peform the actual fixing action (i.e update or delete)
}
\examples{
# Read in FCS files with inconsistencies
fcs_files <- list.files(system.file("extdata", "GvHD_QC", package = "cytoqc"), full.names = TRUE)
qc_cf_list <- cqc_load_fcs(fcs_files)

# Check for marker inconsitencies
groups <- cqc_check(qc_cf_list, type = "marker")

# Attempt to fix them automatically
match_result <- cqc_match(groups, ref = c("CD14 PerCP", "CD15 FITC", "CD33 APC", "CD45 PE", "FSC-Height", "SSC-Height", "Time"))

# Add a manual match that automatic matching could not find
match_result <- cqc_match_update(match_result, map = c("PTPRC PE" = "CD45 PE"))

# Apply the fix to the original cytoframes
cqc_fix(match_result)
}
