% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_io.R
\name{cqc_gs}
\alias{cqc_gs}
\title{Construct a 'cqc_gs' object from a 'GatingSet'}
\usage{
cqc_gs(x)
}
\arguments{
\item{x}{a GatingSet object}
}
\description{
This is mainly for distpatch and wrapping many of the
qc operations on a \code{\link{cqc_cf_list}} object for the GatingSet's
underlying data
}
\examples{
gs_path <- system.file("extdata", "gslist_manual_QC", "gs1", package = "cytoqc")
gs <- load_gs(gs_path)
qc_gs <- cqc_gs(gs)
}
