% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_io.R
\name{cqc_gs_list}
\alias{cqc_gs_list}
\title{Construct a 'cqc_gs_list' object from a list of 'GatingSet' objects}
\usage{
cqc_gs_list(x)
}
\arguments{
\item{x}{a list of 'GatingSet' objects}
}
\description{
For the methods dispatching purpose
}
\examples{
gs_paths <- list.files(system.file("extdata", "gslist_manual_QC", package = "cytoqc"), full.names = TRUE)
gs1 <- load_gs(gs_paths[[1]])
gs2 <- load_gs(gs_paths[[2]])
qc_gs_list <- cqc_gs_list(list(gs1, gs2))
groups <- cqc_check(qc_gslist, type="gate")

}
