% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_io.R
\name{cqc_load_fcs}
\alias{cqc_load_fcs}
\title{Load FCS files
load fcs into 'cqc_cf_list' object which is a list of cytoframes}
\usage{
cqc_load_fcs(files, is_h5 = TRUE, ...)
}
\description{
Load FCS files
load fcs into 'cqc_cf_list' object which is a list of cytoframes
}
