% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_match.R
\name{cqc_match}
\alias{cqc_match}
\title{find the the difference between the reference and target group}
\usage{
cqc_match(x, ...)
}
\arguments{
\item{x}{\code{cqc_check} result returned by cqc_check call}

\item{...}{ref -- specifies the reference, which can be either an integer group id or a character vector giving the actual values of the reference

       select -- the group ids selected for processing

       type -- the qc type (either "channel", "marker", "gate"), automatically determined by the type of \code{x}
       
       max.distance -- Maximum distance allowed for a match. This is passed to the max.distance argument in \code{\link{agrep}}.
       
       partial whether -- to do the partial sub string matching before the approximate string matching}
}
\description{
find the the difference between the reference and target group
}
\examples{
fcs_files <- list.files(system.file("extdata", "GvHD_QC", package = "cytoqc"), full.names = TRUE)
qc_cf_list <- cqc_load_fcs(fcs_files)
channel_groups <- cqc_check(qc_cf_list, type = "channel")
summary(channel_groups)

# Match to a reference group of samples
channel_match <- cqc_match(channel_groups, 3)

# Match to a character vector of target channel names
channel_match <- cqc_match(channel_groups, ref = c("FL1-H", "FL2-A", "FL2-H", "FL3-H", "FL4-H", "FSC-Height", "SSC-Height", "Time"))
}
