% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_match.R
\name{cqc_match}
\alias{cqc_match}
\title{find the the difference between the reference and target group}
\usage{
cqc_match(x, ...)
}
\arguments{
\item{x}{\code{cqc_check} result returned by cqc_check call}

\item{...}{ref specifies the reference, which can be either an integer group id or a characte vector giving the actual values of the reference
select the group ids selected for processing
type the qc type (either "channle", "marker", "gate")}
}
\description{
find the the difference between the reference and target group
}
