% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_group.R
\name{cqc_set_reference}
\alias{cqc_set_reference}
\title{Set a reference sample for further QC.}
\usage{
cqc_set_reference(x, ref)
}
\arguments{
\item{x}{cqc report generated by 'cqc_group'}

\item{ref}{specifies the reference, which can be either an integer group id or a characte vector giving the actual values of the reference}
}
\value{
the original cqc report with the reference info attached
}
\description{
This is the first step prior to finding a QC solution for a data set.
}
