% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_fix.R
\name{cqc_update}
\alias{cqc_update}
\title{Update methods for cyto data}
\usage{
cqc_update(x, from, to, type, ...)
}
\arguments{
\item{x}{cytoframe or GatingSet}

\item{from}{the old value to be updated}

\item{to}{the old value to be updated. Alternatively when type is "marker" it can be a named character which provides
a pair of channel vs marker and 'from' is ignored , instead 'channel' is used as the reference to select target entry to update}

\item{type}{one of qc task "channel", "marker", "keyword", "gate"}

\item{...}{unused}
}
\description{
It is typically called automatically by cqc_fix call
}
