% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_match.R
\name{cqc_update_match}
\alias{cqc_update_match}
\alias{cqc_delete_match}
\title{Manual update/delete the match report}
\usage{
cqc_update_match(x, map, group_id = NULL)

cqc_delete_match(x, map, group_id = NULL)
}
\arguments{
\item{x}{'cqc_match_result_and_solution' object generated by "cqc_match"}

\item{map}{For update, it is a named vector that provides the paired the strings(i.e. old value vs new value)
for deletion, it is a character vector that provides the items to be deleted from matched pairs}

\item{group_id}{the group to update. If NULL, then operate on all groups}
}
\description{
There are cases where the automatic string match from `cqc_match` can't resolve.
This function provides the remedy to manually update the match result.
}
\examples{
\dontrun{
groups <- cqc_check(cqc_data, "marker")
match_result <- cqc_match(groups, ref = 3)
#add the matching pairs to the auto-matching results
match_result <- cqc_update_match(match_result,  map = c("CD197" = "CCR7"), group_id = 1)
#or ommit group_id to apply to all groups
match_result <- cqc_update_match(match_result,  map = c("CD197" = "CCR7"))
#to delete
match_result <- cqc_delete_match(match_result,  "CD197")
}
}
