% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_match.R
\name{cqc_update_match}
\alias{cqc_update_match}
\title{Manual update the match report}
\usage{
cqc_update_match(x, group_id = NULL, map)
}
\arguments{
\item{x}{'cqc_match_result_and_solution' object generated by "cqc_match"}

\item{group_id}{the group to update. If NULL, then operate on all groups}

\item{map}{a named vector that provides the paired the strings(i.e. old value vs new value)}
}
\description{
There are cases where the automatic string match from `cqc_match` can't resolve.
This function provides the remedy to manually update the match result.
}
