% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_check.R
\name{diff.cqc_check}
\alias{diff.cqc_check}
\title{Display the differences among QC groups}
\usage{
\method{diff}{cqc_check}(x, vars, ...)
}
\arguments{
\item{x}{qc table returned by \code{\link{cqc_check}}}

\item{vars}{variable to split by. Determined automatically.}

\item{...}{Additional arguments not for the user. Ignore.}
}
\description{
Display the differences among QC groups
}
\examples{
fcs_files <- list.files(system.file("extdata", "GvHD_QC", package = "cytoqc"), full.names = TRUE)
qc_cf_list <- cqc_load_fcs(fcs_files)

channel_groups <- cqc_check(qc_cf_list, type = "channel")
diff(channel_groups)
}
