% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_params.R
\name{diff.cqc_group}
\alias{diff.cqc_group}
\title{Helper function to only show the difference among qc group}
\usage{
\method{diff}{cqc_group}(x, vars)
}
\arguments{
\item{x}{the grouped summary report generated by 'summary' call on the 'cqc_check' results}
}
\description{
Helper function to only show the difference among qc group
}
\examples{
\dontrun{
su <- summary(groups)
diff(su)

}
}
