% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{knit_print.cqc_group_summary}
\alias{knit_print.cqc_group_summary}
\title{#' @export
print.cqc_group_summary <- function(x, collapse = TRUE, ...){
  if(collapse)
    x <- collapse_params(x)
  class(x) <- class(x)[-1]
  print(x)
}}
\usage{
\method{knit_print}{cqc_group_summary}(x, collapse = TRUE, ...)
}
\description{
#' @export
print.cqc_group_summary <- function(x, collapse = TRUE, ...){
  if(collapse)
    x <- collapse_params(x)
  class(x) <- class(x)[-1]
  print(x)
}
}
