% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_check.R
\name{plot_diff}
\alias{plot_diff}
\title{visualize the tree structure differnece among the GatingSets}
\usage{
plot_diff(groups)
}
\arguments{
\item{groups}{\code{cqc_check_gate} grouping resulte from \code{cqc_check}.}
}
\description{
visualize the tree structure differnece among the GatingSets
}
