% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqc_check.R
\name{summary.cqc_check}
\alias{summary.cqc_check}
\title{Provide the summary view of the cqc_check report.}
\usage{
\method{summary}{cqc_check}(object, ...)
}
\arguments{
\item{object}{qc table returned by \code{\link{cqc_check}}}

\item{...}{Additional arguments not for the user. Ignore.}
}
\description{
It summarize the itemized check report into group overview.
It is automatically triggered by print method of `cqc_check` result.
}
\examples{
fcs_files <- list.files(system.file("extdata", "GvHD_QC", package = "cytoqc"), full.names = TRUE)
cqc_cf_list <- cqc_load_fcs(fcs_files)

channel_groups <- cqc_check(cqc_cf_list, type = "channel")
summary(channel_groups)

}
